server/python/requirements.txt
anyio==4.3.0
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Session==0.6.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
Werkzeug==3.0.1
